Fix discrepency in pc.KeyboardEvent documentation (#1061)
In the docs for pc.KeyboardEvent for the `event` param, it is said that it is 'the original browser event that was fired' but the type assigned for it in the docs is `pc.KeyboardEvent`. Unless I'm mistaken, it should be the `KeyboardEvent` object from the web API. This would correspond to the `event` property which has the same description but is typed `KeyboardEvent`.